# api/openapi.json
{
  "components": {
    "schemas": {
...
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "title": "Detail",
            "type": "array"
          }
        },
        "title": "HTTPValidationError",
        "type": "object"
      },
      "LocationDetail": {
        "description": "One location, with what is kept in it and the locations inside it, by tag.\nDirectly in it, as its page lists them: the box's contents, not the bag's in\nthe box -- those are one GET further.",
        "properties": {
          "asset_id": {
            "title": "Asset Id",
            "type": "string"
          },
          "kind": {
            "title": "Kind",
            "type": "string"
          },
          "locations": {
            "default": [],
            "items": {
              "type": "string"
            },
            "title": "Locations",
            "type": "array"
          },
          "name": {
            "title": "Name",
            "type": "string"
          },
          "notes": {
            "default": "",
            "title": "Notes",
            "type": "string"
          },
          "parent": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Parent"
          },
          "path": {
            "default": "",
            "title": "Path",
            "type": "string"
          },
          "things": {
            "default": [],
            "items": {
              "type": "string"
            },
            "title": "Things",
            "type": "array"
          }
        },
        "required": [
          "asset_id",
          "name",
          "kind"
        ],
        "title": "LocationDetail",
        "type": "object"
      },
      "LocationIn": {
        "description": "A location as it arrives. Every field optional, so a PATCH can name one; a\ncreate without a name is refused by the route rather than here, because a\nPATCH that leaves the name alone is the ordinary case.",
        "properties": {
          "kind": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Kind"
          },
          "name": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Name"
          },
          "notes": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Notes"
          },
          "parent": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Parent"
          }
        },
        "title": "LocationIn",
        "type": "object"
      },
      "LocationOut": {
        "properties": {
          "asset_id": {
            "title": "Asset Id",
            "type": "string"
          },
          "kind": {
            "title": "Kind",
            "type": "string"
          },
          "name": {
            "title": "Name",
            "type": "string"
          },
          "notes": {
            "default": "",
            "title": "Notes",
            "type": "string"
          },
          "parent": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Parent"
          },
          "path": {
            "default": "",
            "title": "Path",
            "type": "string"
          }
        },
        "required": [
          "asset_id",
          "name",
          "kind"
        ],
        "title": "LocationOut",
        "type": "object"
      },
...
      "ValidationError": {
        "properties": {
          "ctx": {
            "title": "Context",
            "type": "object"
          },
          "input": {
            "title": "Input"
          },
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "title": "Location",
            "type": "array"
          },
          "msg": {
            "title": "Message",
            "type": "string"
          },
          "type": {
            "title": "Error Type",
            "type": "string"
          }
        },
        "required": [
          "loc",
          "msg",
          "type"
        ],
        "title": "ValidationError",
        "type": "object"
      }
    }
  },
  "info": {
    "title": "Retro Hardware Database API",
    "version": "0.1.0"
  },
  "openapi": "3.1.0",
  "paths": {
...
    "/api/locations": {
      "get": {
        "operationId": "api_list_locations_api_locations_get",
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "items": {
                    "$ref": "#/components/schemas/LocationOut"
                  },
                  "title": "Response Api List Locations Api Locations Get",
                  "type": "array"
                }
              }
            },
            "description": "Successful Response"
          }
        },
        "summary": "Api List Locations",
        "tags": [
          "locations"
        ]
      },
      "post": {
        "operationId": "api_create_location_api_locations_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/LocationIn"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/LocationOut"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Api Create Location",
        "tags": [
          "locations"
        ]
      }
    },
    "/api/locations/{aid}": {
      "delete": {
        "description": "Delete a location, which is refused while anything is still in it: a thing,\nor a location inside it. Its history goes with it; the moves of the things\nthat were once in it stay on their own histories.",
        "operationId": "api_delete_location_api_locations__aid__delete",
        "parameters": [
          {
            "in": "path",
            "name": "aid",
            "required": true,
            "schema": {
              "title": "Aid",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {}
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Api Delete Location",
        "tags": [
          "locations"
        ]
      },
      "get": {
        "operationId": "api_get_location_api_locations__aid__get",
        "parameters": [
          {
            "in": "path",
            "name": "aid",
            "required": true,
            "schema": {
              "title": "Aid",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/LocationDetail"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Api Get Location",
        "tags": [
          "locations"
        ]
      },
      "patch": {
        "operationId": "api_update_location_api_locations__aid__patch",
        "parameters": [
          {
            "in": "path",
            "name": "aid",
            "required": true,
            "schema": {
              "title": "Aid",
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/LocationIn"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/LocationOut"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Api Update Location",
        "tags": [
          "locations"
        ]
      }
    },
...
  }
}